refactor(library): move loan business logic from model.py to services.py
